+++ src/jsMain/kotlin/at/posselt/pfrpg2e/kingdom/Kingdom.kt
@@ -3,6 +3,7 @@ package at.posselt.pfrpg2e.kingdom
 import at.posselt.pfrpg2e.utils.getAppFlag
 import at.posselt.pfrpg2e.utils.setAppFlag
 import at.posselt.pfrpg2e.utils.unsetAppFlag
+import com.foundryvtt.core.documents.Actor
 import com.foundryvtt.core.utils.deepClone
 import com.foundryvtt.pf2e.actor.PF2EParty
 
@@ -20,3 +21,5 @@ suspend fun KingdomActor.clearKingdom() {
     unsetAppFlag("kingdom-sheet")
 }
 
+fun Actor.isKingdomActor() =
+    this is KingdomActor && getKingdom() != null